LayoutTests/editing/deleting/delete-br-011-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of BODY > HTML > #document to 4 of BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldDeleteDOMRange:range from 0 of DIV > BODY > HTML > #document to 2 of BODY > HTML > #document
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 2 of BODY > HTML > #document to 2 of BODY > HTML > #document toDOMRange:range from 0 of DIV > BODY > HTML > #document to 0 of DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
rdar://problem/4065343 Mail: Deleting a line sometimes makes the insertion point jump to the top of the message.

before deletion:
| "hello"
| <div>
|   class="editing"
|   id="test"
|   <#selection-caret>
|   <br>
| <br>
| "
"

after deletion:
| "hello"
| <div>
|   class="editing"
|   id="test"
|   <#selection-caret>
|   <br>
| "
"
